Logic change to redirect unauthorised access to main menu
